Use Effect system for combat

+++ rules/src/magic.rs
@@ -1,5 +1,5 @@
 /// Various parameters for magic
 #[derive(Serialize, Deserialize)]
 pub struct Magic {
-    pub magic_device_base_power: f64,
+    pub magic_device_base_power: f32,
 }
